refactor: replace Calibration/subject_details
