+++ examples/fip_behavior_instrument.py
@@ -6,8 +6,8 @@ from datetime import date, datetime, timezone
 from aind_data_schema_models.modalities import Modality
 from aind_data_schema_models.units import FrequencyUnit, SizeUnit
 
+from aind_data_schema.components.calibrations import Calibration
 from aind_data_schema.components.devices import (
-    Calibration,
     CameraAssembly,
     CameraTarget,
     Camera,
