+++ src/aind_data_schema/core/acquisition.py
@@ -8,13 +8,14 @@ from pydantic import Field, SkipValidation, model_validator
 from aind_data_schema.base import DataCoreModel, DataModel, AwareDatetimeWithDefault, GenericModel, GenericModelType
 from aind_data_schema_models.units import VolumeUnit, MassUnit
 from aind_data_schema.components.devices import (
-    Calibration,
     Maintenance,
     Camera,
     CameraAssembly,
     EphysAssembly,
     FiberAssembly,
 )
+from aind_data_schema.components.calibrations import Calibration
+
 from aind_data_schema.core.procedures import Anaesthetic
 from aind_data_schema.components.identifiers import Person, Software, Code
